Move App\DependencyInjection\SlaveConfiguration to App\Slave\Configuration
